Workaround "helm home" bug
On occasion, the call to "helm home" is returning repository data instead of
the location to the helm installation.  In such cases, further calls that
explicitly set the "--home" option are incorrect and helmfile fails.  This
attempts to recognize such failures and sets the helm_home variable to a
sane default.